feat(gui): split mask canvas into its own component
